Adds a general framework for merkhets and loggers
As the DOM will be able to run different types of test during the
maintance period, this implements an OO approach to the different task
types. Defined are interfaces for the tasks themselves as well as for
the statistics they will produce at the end of the dom workflow

Introduces Merkhet structure

This commit renames "hurtles" to merkhets as well as restructures the
project files.

It implements the interfaces for the logger pipeline that will be used
to concat the two output streams DOM has to deal with.

Unit tests for the basic merkhet structure were added

Renamed DependencyProvider to Base

The term DependencyProvider seems quite undescriptive and focuses on
OOP. The Base interface now simply holds the base all merkhet instances
will use, and it is a low shorteri in the end as well

Adds interface for Logger coupling

As we will work with different loggers, from the executed commands and
the merkhets, we need a coupler that combines the different loggers into
one final console output.

This also add a GetName method to the Logger, as we may need to provide
that later.

A CouplerFormatter is also being added, which allows to modify
formatting without switching the Coupler implementation

Finalizes the Logger infrastructure DOM uses

This commit finalizese the LoggerCluster which send the messages to the
final loutput

Reformmatted Code to follow standart go structure

The code now follows the default code structure found under
https://github.com/golang-standards/project-layout

It also removes the Base method from the Merkhet instances as they now
provide their Logger and Configuration instances on their own

Fixes multithreaded logger instances

As go channels have the interesting property to block when both the send
and read call are in the same routine, and no dynamicly increasing
buffered go channel exists, this commit introduces a hybrid.

The LoggerChannelProvider now requests a size that will be
configurable. If the buffered channel contains less objects than
possible without blocking the code, the write call is executed in sync.
Any other calls are moved into a new go routine.

Addes logger cluster and pipeline

The logger cluster interface and implementation describe a cluster of
loggers that can be group over a range of differing go-routines.

The cluster is able to listen to the loggers registered on the same
ChannelProvider instance, and groups their incoming messages.

Messages are flushed to the pipeline after a given amount of time, which
is passed to the cluster constructor, or if a one logger tries to write
a second message into the unflushed cluster.

The pipeline as of right now is only an interface, but will later on be
the connection to the terminal, and is responsible for formatting the
intput

Introduces MerkhetWorkerPool and Workers

As merkhets will have to run on different go routines each, this commit
introduces a worker pool that is responsible for controlling them.

The worker instances are connected to the pool through a command channel
that can be used to send requests against the running worker.

For now every new merkhet that is added to the pool, instantly gets a
new running worker node assgined.

Adaptes to go conventions

As requested by QA, this commits fixes issues in the naming conventions.

It removes the java like syntax for getter methods (get/is)
It also fixes a typo found in the source code.

Updates the comment format to include a space after the initial //

Constructor methods now also return the a pointer to the actual type
rather than the inteface pointer to the created struct

Godocs concerning functions declared in interfaces are being moved to
the documentation of the interface itself rather than the methode
declaration

Improves cluster flush logic

The cluster flusg logic can be reduced to a simple if-else part, as the
cluster now handles the timer as a pointer, so we can store a non
initialized timer as nil.

This allows us to not flush directly on the firt message, as the timer
will only start when the listener is enabled

Changes ChannelMessage to value struct

This commit changes the ChannelMessage to be a simply struct using value
recievers, as it will be the main object to communicate between routines
and should be as safe as possible