Adds a simple install script usable from shell
The script downloads the last tagged master commit and installs it into
the usr/local/bin or the ~/bin directory respectivly.

This tool allows for a quick install as it bypasses the need to compile
it on the local machine. This feature is mostly needed in pipelines that
do not want to install golang into their containers

The script is basically a one to one copy from the Havener project:
https://github.com/homeport/havener