Adds interface for Logger coupling
As we will work with different loggers, from the executed commands and
the merkhets, we need a coupler that combines the different loggers into
one final console output.

This also add a GetName method to the Logger, as we may need to provide
that later.

A CouplerFormatter is also being added, which allows to modify
formatting without switching the Coupler implementation