Adds example config.json
This commit adds an example configuration json file that contains the
entire config the DOM tool will need later on.

It does not contain any config parsing code nor provide go interfaces
for those configs.

Comapred to uptimer, this config will allow for either a percentual or
flat treshhold for the merkhets as well as introducing some logger
configuration concerning dates.

It also adds a parameter for custom parameters that will be passed to
the cloud foundery CLI tool, if those are necessary in the future