Fixes multithreaded logger instances
As go channels have the interesting property to block when both the send
and read call are in the same routine, and no dynamicly increasing
buffered go channel exists, this commit introduces a hybrid.

The LoggerChannelProvider now requests a size that will be
configurable. If the buffered channel contains less objects than
possible without blocking the code, the write call is executed in sync.
Any other calls are moved into a new go routine.