Addes logger cluster and pipeline
The logger cluster interface and implementation describe a cluster of
loggers that can be group over a range of differing go-routines.

The cluster is able to listen to the loggers registered on the same
ChannelProvider instance, and groups their incoming messages.

Messages are flushed to the pipeline after a given amount of time, which
is passed to the cluster constructor, or if a one logger tries to write
a second message into the unflushed cluster.

The pipeline as of right now is only an interface, but will later on be
the connection to the terminal, and is responsible for formatting the
intput