Introduces MerkhetWorkerPool and Workers
As merkhets will have to run on different go routines each, this commit
introduces a worker pool that is responsible for controlling them.

The worker instances are connected to the pool through a command channel
that can be used to send requests against the running worker.

For now every new merkhet that is added to the pool, instantly gets a
new running worker node assgined.