Adaptes to go conventions
As requested by QA, this commits fixes issues in the naming conventions.

It removes the java like syntax for getter methods (get/is)
It also fixes a typo found in the source code.

Updates the comment format to include a space after the initial //

Constructor methods now also return the a pointer to the actual type
rather than the inteface pointer to the created struct

Godocs concerning functions declared in interfaces are being moved to
the documentation of the interface itself rather than the methode
declaration