Introduces the configuration code model and parser
This commit adds the golang model for the DOM Configuration.
The parser this commit adds can parse a YAML or a JSON based config, as
it will convert YAML int JSON internally.

This means that the golang model only need the json field tags to be
filled by the unmarshaller.