Improves the logging structure
This commit splits the logger implementation into the logger itself
which now provides a writer instance rather than implementing it.

This change allows one logger instance to create multiple logger
"child" writers that all report to the same logger instance.

A use case of this would be cmd.Stdout = logger.ReportingTo(Info)
This would redirect every output on the stdout channel to the master
logger with the log level info