Adds a simple sample app written in go (#37)
Add a simple sample app that provides a HTTP service as
well as logs to the standart output stream every second

Add a small framework for mapping static assets to
variables in the final go build.

By adding the varible path as well as the fetch command for the asset to
the mappings.properties file, the build script automatically links the
two in the final build.

Introduce `pina-golada` to both the asset implementation as
well as the build file to automatically generate the files needed

Updates all file formatting and introduces
a space after each double slash comment.

Re-appliy the branch onto the develop branch. It also contains
a few typo fixes.