Introduces merkhet heartbeat
The merkhet workers have to be ticked by a heart. This commit introduces
a new wrapper around the Worker instances that, once started, will
executed a pre-set consumer every x time units.

This heartbeat can be started independently of the worker instances
themselves, to allow a pre-heartbeat setup call.
Heartbeats will be shutdown with the workers on pool#shutdown