Implements general workflow of watchful
This commit adds create/delete org/space methods to the CloudFoundryCLI
interface/implementation. Those will be needed to create the test
organization watchful will run in

Implements the CloudFoundryWorker

This commit implements the CloudFoundryWorker as the logger system is
now able to log both stdout and stderr correctly

Reworked logger group config

This commit reworks the logger group config and prepares a general run
command. The project can now create a basic cloud foundry env for
testing

Adds terminal width as parameter to run command

As the terminal width may not always be fetchable, this parameter allows
to override it

Implements a first WIP general workflow

This commit implements a first general workflow of the run command as
well as provides config reading / general parameters using cobra

Now runs the tasks configured in the config

This commit allows the general workflow to run the tasks configured in the config. Merkhets are not implemented yet, but the main update task is working

Fixes time location usage

This commit actually makes the logger pipeline adapt the found time
instance to the provided time-zone

Improves the error forwarding of the cf worker

The cloud foundry worker now forwards custom signal implementations to
the main watchful service to provide more detailed logs at the end of
the program