Prints the entire merkhet result log
This commit now waits on every merkhet result task to finish before
terminating the routine itself. This will therefore print every merkhet
result not just the first failing.

This commit also fixes the issue of merkhets starting when the pool is
about to shutdown, as the pool would wait for all merkhets to finish
their current task while not closing the heartbeats prior to that which
lead to an endless loop in the worst case